requirements.txt
Package             Version
------------------- -------
asgiref             3.2.10
Django              3.1.1
djangorestframework 3.11.1
mysqlclient         2.0.1
pip                 19.0.3
pytz                2020.1
setuptools          40.8.0
sqlparse            0.3.1
